+++ changelog.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.2.19] - 2026-09-25
 
+## Fixed
+
+- Transform `set_fill`/`set_stroke` parameter from handling "CSS colour value" to "SVG paint value" (``)
+
 ## Changed
 
 - Doc only: Refactored dcos `.md` files (`1b77086`)